+++ libs/tendermint/rpc/jsonrpc/client/converters.go
@@ -2,10 +2,29 @@ package client
 
 import (
 	abci "github.com/okex/exchain/libs/tendermint/abci/types"
+	"github.com/okex/exchain/libs/tendermint/libs/bytes"
 	coretypes "github.com/okex/exchain/libs/tendermint/rpc/core/types"
+	"github.com/okex/exchain/libs/tendermint/types"
 )
 
-func ConvCM39ToCM4(c *coretypes.CM39ResultTx, ret *coretypes.ResultTx) {
+// Query abci msg
+type CM39ResultABCIQuery struct {
+	Response abci.CM39ResponseQuery `json:"response"`
+}
+
+// Result of querying for a tx
+type CM39ResultTx struct {
+	Hash     bytes.HexBytes             `json:"hash"`
+	Height   int64                      `json:"height"`
+	Index    uint32                     `json:"index"`
+	TxResult abci.CM39ResponseDeliverTx `json:"tx_result"`
+	Tx       types.Tx                   `json:"tx"`
+	Proof    types.TxProof              `json:"proof,omitempty"`
+}
+
+/////////
+
+func ConvCM39ToCM4(c *CM39ResultTx, ret *coretypes.ResultTx) {
 	ret.Hash = c.Hash
 	ret.Height = c.Height
 	ret.Index = c.Index
@@ -22,7 +41,7 @@ func ConvCM39ToCM4(c *coretypes.CM39ResultTx, ret *coretypes.ResultTx) {
 	ret.Proof = c.Proof
 }
 
-func ConvTCM39ResultABCIQuery2CM4(c *coretypes.CM39ResultABCIQuery, ret *coretypes.ResultABCIQuery) {
+func ConvTCM39ResultABCIQuery2CM4(c *CM39ResultABCIQuery, ret *coretypes.ResultABCIQuery) {
 	ret.Response = abci.ResponseQuery{
 		Code:      c.Response.Code,
 		Log:       c.Response.Log,
